WIP: migration to CMake. o Creating header for msgflo. o Upgrading to latest release of amqpcpp. o Building amqpcpp with CMake too. o Making the example project a standalone CMake project so it tests the entire CMake setup.
