+++ src/msgflo.cpp
@@ -1,4 +1,3 @@
-
 // cd thirdparty/amqpcpp && make install PREFIX=./install; cd -
 // g++ -std=c++11 -o ./spec/fixtures/repeat-cpp ./spec/fixtures/repeat.cpp -I./thirdparty/json11/ -I./thirdparty/amqpcpp/install/include/ -I./thirdparty/amqpcpp/examples/rabbitmq_tutorials/  ./thirdparty/amqpcpp/install/lib/libamqpcpp.a.2.2.0 -lboost_system -pthread
 
@@ -7,181 +6,42 @@
 #include <thread>
 #include <chrono>
 #include <boost/algorithm/string.hpp>
-
-#include "json11.cpp"
-#include "asiohandler.cpp"
+#include <boost/asio/io_service.hpp>
+#include "amqpcpp.h"
+#include "json11.hpp"
+#include "msgflo.h"
 
 namespace msgflo {
+
+using std::string;
 
 class Engine;
 class Participant;
 
-struct Definition {
-
-    struct Port {
-        std::string id;
-        std::string type;
-        std::string queue;
-
-        json11::Json to_json() const
-        {
-            return json11::Json::object {
-                { "id", id },
-                { "type", type },
-                { "queue", queue }
-            };
-        }
-    };
-
-public:
-    static std::string queueForPort(const std::vector<Port> &ports, std::string id) {
-        for (const auto &p : ports) {
-            if (p.id == id) {
-                return p.queue;
-            }
-        }
-        return "";
+class MsgFloAmqpHandler : public AMQP::TcpHandler {
+    virtual void onConnected(AMQP::TcpConnection *connection) {
     }
 
-    static Port &addDefaultQueue(Port &p, std::string role) {
-        if (!p.queue.empty()) {
-            return p;
-        }
-        p.queue = role+"."+boost::to_upper_copy<std::string>(p.id);
-        return p;
+    virtual void onError(AMQP::TcpConnection *connection, const char *message) {
     }
 
-    static Definition instantiate(Definition &d, std::string role)
-    {
-        d.id = role + std::to_string(rand());
-        d.role = role;
+    virtual void onClosed(AMQP::TcpConnection *connection) {}
 
-        for (auto &p : d.inports) {
-            addDefaultQueue(p, role);
-        }
-        for (auto &p : d.outports) {
-            addDefaultQueue(p, role);
-        }
-        return d;
+    virtual void monitor(AMQP::TcpConnection *connection, int fd, int flags) {
     }
-
-public:
-    Definition() 
-        : inports {
-            Port{ "in", "any", "" }
-        },
-        outports {
-            Port{ "out", "any", "" },
-            Port{ "error", "error", "" }
-        }
-    {
-    }
-
-    json11::Json ports_to_json(const std::vector<Port> &ports) const
-    {
-        using namespace json11;
-        auto objects = std::vector<Json>();
-        for (const auto &p : ports) {
-            objects.push_back(p.to_json());
-        }
-        return Json::array(objects);
-    }
-
-    json11::Json to_json() const
-    {
-        using namespace json11;
-
-        auto ins = ports_to_json(inports);
-        return Json::object {
-            { "id", id },
-            { "role", role },
-            { "component", component },
-            { "label", label },
-            { "icon", icon },
-            { "inports", ins },
-            { "outports", ports_to_json(outports) }
-        };
-    }
-
-public:
-    std::string id;
-    std::string role;
-    std::string component;
-    std::string label = "";
-    std::string icon = "file-word-o";
-    std::vector<Port> inports;
-    std::vector<Port> outports;
-};
-
-
-struct Message {
-    // TODO: support binary type also
-    enum Type {
-        Json,
-        Binary
-    };
-
-    uint64_t deliveryTag;
-    json11::Json json;
-    Type type = Json;
-};
-
-class Participant {
-    friend class Engine;
-
-public:
-    Participant(std::string _role, Definition _def)
-        : role(_role)
-    {
-        definition = Definition::instantiate(_def, role);
-    }
-
-protected:
-    /* Receiving messages. Override in subclass */
-    virtual void process(std::string port, Message msg) = 0;
-
-    /* Sending messages */
-    void send(std::string port, Message &msg);
-
-    // ACK/NACK
-    void ack(Message msg);
-    void nack(Message msg);
-
-private:
-    std::string role;
-    Engine* _engine = nullptr;
-    Definition definition;
-};
-
-
-struct ConnectionOptions {
-    ConnectionOptions()
-        : port(5672)
-        , host("localhost")
-        , user("guest")
-        , password("guest")
-    {
-    }
-
-    // TODO: add code for parsing from string
-
-    int port;
-    std::string host;
-    std::string user;
-    std::string password;
 };
 
 class Engine {
     friend class Participant;
 
 public:
-    Engine(Participant *p, ConnectionOptions o)
-        : handler(ioService)
-        , connection(&handler, AMQP::Login(o.user, o.password), "/")
+    Engine(Participant *p, const string &url)
+        : handler()
+        , connection(&handler, AMQP::Address(""))
         , channel(&connection)
         , participant(p)
     {
-        handler.connect(o.host, o.port);
+//        handler.connect(o.host, o.port);
         channel.setQos(1); // TODO: is this prefech?
         participant->_engine = this;
     }
@@ -201,7 +61,7 @@ public:
     }
 
     void stop() {
-        
+
     }
 
 private:
@@ -238,7 +98,7 @@ private:
     // Interfaces used by Participant
     void send(std::string port, const Message &msg) {
         const std::string exchange = Definition::queueForPort(participant->definition.outports, port);
-        const std::string data = msg.json.dump(); 
+        const std::string data = msg.json.dump();
         AMQP::Envelope env(data);
         std::cout <<" Sending on " << exchange << " :" << data << std::endl;
         channel.publish(exchange, "", env);
@@ -254,12 +114,11 @@ private:
 
 private:
     boost::asio::io_service ioService;
-    AsioHandler handler;
-    AMQP::Connection connection;
-    AMQP::Channel channel;
+    MsgFloAmqpHandler handler;
+    AMQP::TcpConnection connection;
+    AMQP::TcpChannel channel;
     Participant *participant;
 };
-
 
 void Participant::send(std::string port, Message &msg)
 {
@@ -277,7 +136,5 @@ void Participant::nack(Message msg) {
     _engine->nack(msg);
 }
 
-
 } // namespace msgflo;
-
